PDF Will Finished
PDF will finished
Users will not be asked to create a new will after completing one in their user area. Instead it will allow them to download their will

+++ plugins/willwritingpartnership/diywill/updates/create_funeral_arrangements_table.php
@@ -31,29 +31,14 @@ class CreateFuneralArrangementsTable extends Migration
             $table->timestamps();
         });
 
-        Schema::create('will', function(Blueprint $table) {
-            $table->engine = 'InnoDB';
-            $table->increments('id');
-            $table->integer('userid')->nullable();
-            $table->foreign('userid')->references('id')->on('useraccount');
-            $table->integer('octoberid')->nullable();
-            $table->foreign('octoberid')->references('id')->on('users');
-            $table->integer('residestateid')->nullable();
-            $table->foreign('residestateid')->references('id')->on('residualestate');
-            $table->integer('executors')->nullable();
-            $table->foreign('executors')->references('id')->on('appointedexecutors');
-            $table->integer('funeralID')->nullable();
-            $table->foreign('funeralID')->references('id')->on('funeralarrangements');
-            $table->boolean('complete')->default(false);
-            $table->integer('progress')->default(0);
-            $table->boolean('paid')->default(false);
-            $table->timestamps();
-        });
     }
 
     public function down()
     {
-        Schema::dropIfExists('FuneralArrangements');
-        Schema::dropIfExists('will');
+        if(Schema::hasTable('will')){
+            Schema::dropIfExists('will');
+        }
+        Schema::dropIfExists('funeralarrangements');
+
     }
 }
